Initial test update commit. Renamed tests, modified distribution tests to run with numba, modified surface tests to run with pytest.
