+++ pyproject.toml
@@ -95,5 +95,6 @@ force-exclude = '''
   mcdc/numba_types\.py
   | mcdc/mcdc_get/
   | mcdc/mcdc_set/
+  | mcdc/lib/
 )
 '''
