+++ test/unit/transport/distributions/conftest.py
@@ -1,54 +1,66 @@
-import os
+import numpy as np
 import pytest
-
-# Force pure-Python execution for numba in tests.
-os.environ.setdefault("NUMBA_DISABLE_JIT", "1")
+from numba import njit
 
 import mcdc.transport.distribution as dist
 
+MOCK_RNG_MAX_VALUES = 8
+MOCK_RNG_STATE_DTYPE = np.dtype(
+    [
+        ("idx", np.int64),
+        ("n_values", np.int64),
+        ("values", np.float64, (MOCK_RNG_MAX_VALUES,)),
+    ]
+)
 
-class MockRNG:
-    def __init__(self, values):
-        self._values = list(values)
-        self._i = 0
 
-    def lcg(self, _state_container):
-        # MCDC uses dist.rng.lcg(state) as its RNG hook. We override it with this
-        # deterministic sequence so tests can be analytic and reproducible.
-        if self._i >= len(self._values):
-            raise IndexError("MockRNG depleted")
-        value = self._values[self._i]
-        self._i += 1
-        return value
+@njit
+def mock_lcg_known_sequence(rng_state):
+    # Force lcg to pick mock rng values in sequence
+    state = rng_state[0]
+    i = state["idx"]
+    n = state["n_values"]
+
+    if i < n:
+        value = state["values"][i]
+    else:
+        value = np.nan
+
+    state["idx"] = i + 1
+    return value
+
+
+def mock_rng(*values):
+    if len(values) == 1 and isinstance(values[0], (list, tuple, np.ndarray)):
+        values = tuple(values[0])
+
+    state = np.zeros(1, dtype=MOCK_RNG_STATE_DTYPE)
+    state[0]["idx"] = 0
+    state[0]["n_values"] = len(values)
+    if values:
+        state[0]["values"][: len(values)] = np.asarray(values, dtype=np.float64)
+    return state
 
 
 @pytest.fixture
-def rng_state():
-    """
-    Return the minimal RNG state container expected by MCDC.
+def make_distribution_record():
+    # Build one typed record from a dict of field values.
+    def _make(record_dtype, values_dict):
+        container = np.zeros(1, record_dtype)
+        record = container[0]
+        for key, value in values_dict.items():
+            record[key] = value
+        return record
 
-    MCDC passes a list of per-thread state dicts; tests only need one.
-    """
-    return [dict(rng_seed=0)]
+    return _make
 
 
 @pytest.fixture
-def rng_sequence():
-    """
-    Temporarily replace dist.rng.lcg with a deterministic sequence.
-
-    Usage:
-        rng_sequence([0.1, 0.2, 0.3])
-        ... code under test that calls dist.rng.lcg(...)
-    """
+def mock_rng_sequence():
+    # Replace mcdc lcg with the known sequence
     original_lcg = dist.rng.lcg
-
-    def _apply(values):
-        # Swap in a mock LCG implementation that returns the provided values.
-        rng = MockRNG(values)
-        dist.rng.lcg = rng.lcg
-        return rng
-
-    # Yield a callable to the test, then restore the real RNG after the test ends.
-    yield _apply
+    dist.rng.lcg = mock_lcg_known_sequence
+    # Yield the mock_rng sequence
+    yield mock_rng
+    # Reset the lcg. If the test fails, the mock_rng could still be installed and cause later tests to fail
     dist.rng.lcg = original_lcg
